opencsd: Update PE architecture version handling in decoder.
PE architecture version has effect on packet interpretation between
v7/v8 A and M profiles, and on valid instruction opcodes for waypoints in
trace instruction decode.

Beyond v8.3 architecture, PE features can be cherry picked from a later version
to appear on a core design of an earlier architecture version.

To increase flexibility in support for new features, a new generic architecture
value constant is defined as ARCH_AA64, which is defined to mean any PE with a
feature set of v8.3 plus addtional features.

Instruction decode is adjusted to filter using min version macros rather than
direct comparison.

Trace snapshot files will now support device 'type'  strings of form
ARMv8.x-{A|R|M}, plus ARM-AA64.

Test program mapping will set the ARCH_xxx version appropriately.

+++ decoder/source/i_dec/trc_i_decode.cpp
@@ -42,7 +42,7 @@ ocsd_err_t TrcIDecode::DecodeInstruction(ocsd_instr_info *instr_info)
     struct decode_info info;
 
     info.instr_sub_type = OCSD_S_INSTR_NONE;
-    info.arch_version = (uint16_t)(instr_info->pe_type.arch);
+    info.arch_version = instr_info->pe_type.arch;
 
     switch(instr_info->isa)
     {
